Rename quiz model file
