+++ MarvelUniverse/models/__init__.py
@@ -1,4 +1,4 @@
 from .marvel_models import Character, Comic, Series, CharacterInComic, CharacterInSeries
 from .comment_models import SeriesComment, ComicComment, CharacterComment
 from .favorites import FavoriteCharacter, FavoriteComic, FavoriteSeries
-from .quiz import CharacterQuestion, CharacterChoice, SeriesQuestion, SeriesChoice, ComicQuestion, ComicChoice
+from .quiz_models import CharacterQuestion, CharacterChoice, SeriesQuestion, SeriesChoice, ComicQuestion, ComicChoice
